+++ spec/support/integration_helpers.rb
@@ -5,7 +5,6 @@ module IntegrationHelpers
     fill_in "Email",    with: user.email
     fill_in "Password", with: password
     click_button 'Sign in'
-    find(:css, '.card-link', match: :first).click
   end
 
   def send_keys keys, options = {}