+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Redirect automatically to project#index after the user authentication, when there is only one team on enrollments
 
 ## [1.1.3] - 2017-03-30
 ### Changed